+++ crown-court-proceeding/build.gradle
@@ -89,6 +89,7 @@ dependencies {
     testImplementation 'org.testcontainers:localstack:1.18.3'
     //testImplementation 'org.testcontainers:junit-jupiter:1.18.3'
     testImplementation 'cloud.localstack:localstack-utils:0.2.23'
+    testImplementation group: 'com.github.tomakehurst', name: 'wiremock', version: '2.27.2', ext: 'pom'
 
     implementation group: 'com.amazonaws', name: 'aws-java-sdk', version: '1.12.264'
     implementation group: 'com.amazonaws', name: 'aws-java-sdk-sqs', version: '1.12.264'
